Move sor into modules dir
