+++ balancer-js/src/modules/sor/pool-data/subgraphPoolDataService.ts
@@ -8,7 +8,7 @@ import { parseInt } from 'lodash';
 import { getOnChainBalances } from './onChainData';
 import { Provider } from '@ethersproject/providers';
 import { Network } from '@/lib/constants/network';
-import { BalancerNetworkConfig, BalancerSdkSorConfig } from '../../types';
+import { BalancerNetworkConfig, BalancerSdkSorConfig } from '../../../types';
 
 const NETWORKS_WITH_LINEAR_POOLS = [
     Network.MAINNET,
